Move the Maze POMDP family out of the t_maze module names
The Maze family already existed on develop, but it lived in files named after
the T-Maze. `environments/t_maze_pomdp/` held `maze_pomdp.py`, `maze_geometry.py`
and `t_maze_visualizer.py` -- which, despite its name, defines the general
`MazeVisualizer` used by every Maze variant. `environments/maze_pomdp/` was a
single file that forwarded imports back. The general code sat under the special
case's name.

Flip it: `maze_pomdp` is now the real package. `t_maze_pomdp.py` moves with it
and keeps its name, because `TMazePOMDP` really is the T-shaped special case.
`t_maze_visualizer.py` becomes `maze_visualizer.py`, and `TMazeVisualizer` stays
as an alias in the compatibility layer only.

`environments/t_maze_pomdp/` remains importable as forwarding shims rather than
being deleted. `Environment.from_dict` imports the module string recorded in a
saved configuration verbatim, so a config written before this move names a
`t_maze_pomdp` module and would fail to load without them.

Renderer behaviour and the golden GIF are untouched, and all three environment
`config_id` values are unchanged -- they are built from an instance's attributes,
not its module path -- so no cached simulation result is invalidated.
