+++ POMDPPlanners/environments/t_maze_pomdp/__init__.py
@@ -1,43 +1,57 @@
 # SPDX-License-Identifier: MIT
 
-"""Maze POMDP family, including the legacy T-shaped compatibility class.
+"""Compatibility alias for :mod:`POMDPPlanners.environments.maze_pomdp`.
 
-Classes:
-    DiscreteMazePOMDP: Generated maze with cell actions.
-    ContinuousMazePOMDP: The same generated maze with real displacement actions.
-    TMazePOMDP: Compatibility class preserving the original T-shaped geometry.
+The Maze family moved to the ``maze_pomdp`` package, because nothing here was
+T-specific: ``DiscreteMazePOMDP``, ``ContinuousMazePOMDP`` and the shared renderer
+all lived under the special case's name. This package and its submodules stay
+importable because :meth:`Environment.from_dict` imports the module path recorded
+in a saved configuration, so dropping them would break every Maze or T-Maze
+configuration saved before the move.
+
+New code should import from ``POMDPPlanners.environments.maze_pomdp``.
 """
 
-from POMDPPlanners.environments.t_maze_pomdp.maze_geometry import MazeGeometry
-from POMDPPlanners.environments.t_maze_pomdp.maze_pomdp import (
-    BaseMazePOMDP,
-    ContinuousMazePOMDP,
-    DiscreteMazePOMDP,
-    MazeMetric,
-    MazeStepChannel,
-    StepOutcome,
-    create_maze_state,
+# The submodules are imported for their side effect of binding themselves as
+# attributes of this package. Before the move the package imported from its own
+# submodules, so ``t_maze_pomdp.maze_pomdp`` resolved by attribute access and not
+# only by ``import``; keeping that working costs five re-export modules.
+from POMDPPlanners.environments.t_maze_pomdp import (  # noqa: F401
+    maze_geometry,
+    maze_pomdp,
+    maze_visualizer,
+    t_maze_pomdp,
+    t_maze_visualizer,
 )
 
-from POMDPPlanners.environments.t_maze_pomdp.t_maze_pomdp import (
+from POMDPPlanners.environments.maze_pomdp import (  # noqa: F401
     ACTIONS,
+    BaseMazePOMDP,
     CUE_CONSUMED,
     CUE_EMITTING,
     CUE_UNSEEN,
     GOAL_LEFT,
     GOAL_RIGHT,
+    ContinuousMazePOMDP,
+    DiscreteMazePOMDP,
+    MazeGeometry,
+    MazeMetric,
+    MazeStepChannel,
+    MazeVisualizer,
+    OBSERVATIONS,
     OBSERVATION_EMPTY,
     OBSERVATION_LEFT_CUE,
     OBSERVATION_RIGHT_CUE,
-    OBSERVATIONS,
     STATE_CUE_PHASE,
     STATE_GOAL,
     STATE_WIDTH,
     STATE_X,
     STATE_Y,
+    StepOutcome,
     TMazeMetric,
     TMazePOMDP,
     TMazeStepChannel,
+    create_maze_state,
     create_t_maze_state,
 )
 
@@ -54,6 +68,7 @@ __all__ = [
     "MazeGeometry",
     "MazeMetric",
     "MazeStepChannel",
+    "MazeVisualizer",
     "OBSERVATIONS",
     "OBSERVATION_EMPTY",
     "OBSERVATION_LEFT_CUE",
@@ -63,10 +78,10 @@ __all__ = [
     "STATE_WIDTH",
     "STATE_X",
     "STATE_Y",
+    "StepOutcome",
     "TMazeMetric",
     "TMazePOMDP",
     "TMazeStepChannel",
-    "StepOutcome",
     "create_maze_state",
     "create_t_maze_state",
 ]
